Move most of it out again as applyFixes.js
